move the bits and bobs around
