+++ snuba/api/static/config.html
