+++ snuba/api/static/dashboard.html
