+++ snuba/api/templates/base.html
