+++ snuba/cli/api.py
@@ -5,7 +5,7 @@ import click
 @click.option("--debug", is_flag=True)
 def api(*, debug: bool) -> None:
     from snuba import settings
-    from snuba.views import application
+    from snuba.api.views import application
     from werkzeug.serving import WSGIRequestHandler
 
     WSGIRequestHandler.protocol_version = "HTTP/1.1"
